requirements.txt
setuptools
de-nina
shapely
notify_py
python-dotenv
platformdirs
tomli-w
